requirements.txt
Flask==3.1.1
flask-jwt-extended
python-dotenv==1.1.1
PyMySQL==1.1.1
bcrypt==4.3.0